pyproject.toml
[project]
name = "personal-accounts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.5.5",
    "aioboto3>=15.0.0",
    "aiosmtplib>=3.0.2",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.3",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.29.0",
    "autoflake>=2.3.1",
    "bcrypt<4.0.0",
    "black>=24.10.0",
    "environs>=11.0.0",
    "factory-boy>=3.3.1",
    "fastapi-limiter>=0.1.6",
    "fastapi-mail>=1.4.2",
    "fastapi[standard]>=0.115.0",
    "faststream>=0.5.45",
    "flake8>=7.1.1",
    "gunicorn>=23.0.0",
    "isort>=5.13.2",
    "itsdangerous>=2.2.0",
    "openpyxl>=3.1.5",
    "orjson>=3.10.10",
    "passlib>=1.7.4",
    "pre-commit>=4.0.0",
    "pydantic-settings>=2.5.2",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.17",
    "pytz>=2024.2",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.35",
    "starlette-admin>=0.15.1",
    "uvicorn>=0.31.0",
    "websocket-client>=1.8.0",
]

[tool.setuptools]
packages = ["app", "scripts"]

[project.scripts]
dev = "scripts.commands:dev"
migrate = "scripts.commands:migrate"
prod = "scripts.commands:prod"
create_superuser = "scripts.commands:run_create_superuser"
rollback = "scripts.commands:rollback"
serve = "scripts.commands:serve"
start = "scripts.commands:start_all"
test = "scripts.commands:test"

[[tool.mypy.overrides]]
ignore_errors = true
module = "app.core.migrations.*"
